+++ CHANGELOG.md
@@ -21,3 +21,6 @@
   * Added `GanbaroDigital\OperatingSystem\OsType\ValueBuilders\BuildTypeFromEtcRedhatRelease` to use `/etc/redhat-release` to determine which Redhat Linux-based distro we have
   * Added `GanbaroDigital\OperatingSystem\OsType\ValueBuilders\BuildTypeFromLsbRelease` to use the output of `/usr/bin/lsb_release` to determine which Linux distro we are using
   * Added `GanbaroDigital\OperatingSystem\OsType\ValueBuilders\BuildTypeFromSwVers` to use the output of `/usr/bin/sw_vers` to determine which operating system we are using
+* Support for dealing with operating systems we cannot detect
+  * Added `GanbaroDigital\OperatingSystem\OsType\Values\Unknown` for when we do not know what operating system we have
+  * Added `GanbaroDigital\OperatingSystem\OsType\ValueBuilders\EnsureOsType` to return `Unknown` when the value builders do not return a valid `OsType` value